% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR1.R
\name{AR1}
\alias{AR1}
\title{First-order autoregressive model}
\usage{
AR1(x)
}
\arguments{
\item{x}{a vector, ideally a timeseries vector}
}
\value{
an Arima model
}
\description{
This function is a wrapper around the \code{\link[forecast:Arima]{forecast::Arima()}} function.
It creates a first-order autoregressive model with a constant term.
The \link[forecast:Arima]{Arima()} function is itself a wrapper of
the \link[stats:arima]{arima()} function.
}
\examples{
(m1 <- AR1(LakeHuron))
AIC(m1)
(m2 <- AR1(bacon$value))

}
