% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{decompose_branches}
\alias{decompose_branches}
\title{Decompose Branches}
\usage{
decompose_branches(pairs)
}
\arguments{
\item{(pairs)}{Matching representation of the tree (output of matching function)}
}
\value{
A (2n -2)x(n-1) matrix where each row represents a branch and
         each column represents an intercoalescent time. Each branch length b_j
         can be written as a sum of intercoalescent times. A[j, i] = 1 if branch j uses intercoalescent time i.
}
\description{
This function decomposes the branch lengths of the tree into a matrix of intercalescent times.
}
\examples{
plot(mytree)
nodelabels()
#' pairs = matching(mytree)
#      [,1] [,2] [,3]
# [1,]    1    8    9
# [2,]    2    7    8
# [3,]    6    5    7
# [4,]    3    4    6

decompose_branches(pairs)
#     [,1] [,2] [,3] [,4]
# [1,]    1    1    1    1
# [2,]    1    1    1    0
# [3,]    1    0    0    0
# [4,]    1    0    0    0
# [5,]    1    1    0    0
# [6,]    0    1    0    0
# [7,]    0    0    1    0
# [8,]    0    0    0    1
}
