% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_theta.R
\name{flatten_upper}
\alias{flatten_upper}
\title{Flatten Upper Triangular Matrix by Diagonals}
\usage{
flatten_upper(M)
}
\arguments{
\item{M}{A square matrix.}
}
\value{
A numeric vector containing the upper triangular elements of \code{M},
        ordered by diagonal.
}
\description{
Converts the upper triangular part of a square matrix into a vector,
traversing along successive diagonals starting from the main diagonal.
}
\details{
For example, the matrix:
\preformatted{
[theta1, theta12, theta13]
[0     , theta2 , theta23]
[0     , 0      , theta3 ]
}
becomes:
\code{c(theta1, theta2, theta3, theta12, theta23, theta13)}.
}
\examples{
M <- matrix(c(1, 2, 3,
              0, 4, 5,
              0, 0, 6), nrow = 3, byrow = TRUE)
flatten_upper(M)
# Output: 1 4 6 2 5 3

}
