% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello1}
\alias{hello1}
\title{Hello, world!}
\usage{
hello1()
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!'.
}
\details{
You can learn more about package authoring with RStudio at:

https://r-pkgs.org

Some useful keyboard shortcuts for package authoring:

Install Package:           'Cmd + Shift + B'
Check Package:             'Cmd + Shift + E'
Test Package:              'Cmd + Shift + T'
}
