% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lettuce}
\alias{lettuce}
\title{FRED Lettuce data}
\format{
\subsection{\code{lettuce}}{

A data frame with 119 rows and 5 columns:
\describe{
\item{date}{a Date object with the data of the data}
\item{series_id}{APU0000712311, lettuce series}
\item{value}{the average price of iceberg lettuce}
\item{realtime_start, realtime_end}{The date the data was downloaded}
}
}
}
\source{
\url{https://fred.stlouisfed.org/series/APU0000712211}
}
\usage{
lettuce
}
\description{
A subset of the time series APU0000712211 from FRED, data on lettuce
prices from 2016 to 2026
}
\keyword{datasets}
