% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{lumped_state}
\alias{lumped_state}
\title{Lumped State}
\usage{
lumped_state(state)
}
\arguments{
\item{state}{sx1 vector of allele states}
}
\value{
sx1 vector of lumped states using mutation state notation
}
\description{
Removes mutation labels from the full state to give a lumped state
}
\examples{
simul1 <- phylodyn::coalsim(samp_times = 0, n_sampled = 5, traj = exp_traj,method="tt",val_upper=11)
mytree <- phylodyn::generate_newick((simul1))$newick
theta <- matrix(c(1, 1, 1, 0, 1, 1, 0, 0, 1), nrow = 3, byrow = TRUE)
states_matrix <- state_space_matrix(3)
q <- q_matrix(states_matrix, theta)
M <- matrix(20, nrow = 3, ncol = 3)
datalist <- simulate_data_finite_alleles(mytree, q, states_matrix, M)
datalist[[2]]
#      [,1]   [,2]   [,3]
# [1,] "1:18" "0"    "3:7"
# [2,] "0"    "0"    "0"
# [3,] "0"    "2:8"  "0"
# [4,] "0"    "0"    "0"
# [5,] "1:20" "23:8" "23:8"
# [6,] "0"    "0"    "0"
# [7,] "0"    "0"    "0"
# [8,] "0"    "0"    "0"
# [9,] "0"    "0"    "0"

lumped_state(datalist[[2]][1, ])
# 1 0 1
lumped_state(datalist[[2]][5, ])
# 1 3 3

}
