% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{matching}
\alias{matching}
\title{Tree mapping representation}
\usage{
matching(tree)
}
\arguments{
\item{(tree)}{An ape tree object}
}
\value{
(n-1)x3 matrix where each row is a merge event. The first two coordinates
         represent the labels of the children and the third coordinate is the label of the parent node
}
\description{
This function returns the matching representation of an ape tree as a matrix.
Each row is a merge and the third entry in the row is the label of the interior node for th emerge.
Leaves are 1,...,n and interior nodes are n+1, .., 2n-1 (Note that ape labels with te root as n+1)
Warning: this may give invalid tree if an edge length is 0 (which can happen with upgma)
}
\examples{
matching(mytree)
#      [,1] [,2] [,3]
# [1,]    1    8    9
# [2,]    2    7    8
# [3,]    6    5    7
# [4,]    3    4    6

}
