% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{observed_mutations}
\alias{observed_mutations}
\title{Observed Mutations}
\usage{
observed_mutations(d, s, lumped_d = NA)
}
\arguments{
\item{d}{Matrix of allele states at the tips}

\item{s}{Number of sites}

\item{lumped_d}{(Optional) Matrix of lumped states at the tips}
}
\value{
An sxs matrix where s[i,i] is number of cells with a single mutation
         in site i, s[i, j], i < j, is number of cells with an overlapping
         mutation going from i to j
}
\description{
This function counts the number of observed mutations from the matrix D,
which stores alleles data from the tips. Note that this function will return
counts only based on the lumped mutation state, not the allele states
}
\examples{
simul1 <- phylodyn::coalsim(samp_times = 0, n_sampled = 5, traj = exp_traj,method="tt",val_upper=11)
mytree <- phylodyn::generate_newick((simul1))$newick
theta <- matrix(c(1, 1, 1, 0, 1, 1, 0, 0, 1), nrow = 3, byrow = TRUE)
states_matrix <- state_space_matrix(3)
q <- q_matrix(states_matrix, theta)
M <- matrix(20, nrow = 3, ncol = 3)
datalist <- simulate_data_finite_alleles(mytree, q, states_matrix, M)
observed_mutations(datalist[[2]], 3)
# $obsM
#      [,1] [,2] [,3]
# [1,]    2    0    0
# [2,]    0    1    1
# [3,]    0    0    1
#
# $obsA
# [1] 2 1 1 0 1 0
}
