% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_missing.R
\name{perc_missing}
\alias{perc_missing}
\title{Find percent missing of time series}
\usage{
perc_missing(x)
}
\arguments{
\item{x}{a vector, ideally a time series}
}
\value{
the percentage of the time series that is NA, as a character string
}
\description{
A wrapper around \link[imputeTS:statsNA]{imputeTS::statsNA}
}
\examples{
perc_missing(bacon$value)
perc_missing(lettuce$value)
perc_missing(tomatoes$value)
}
