% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{probs_function}
\alias{probs_function}
\title{Probability of Mutations}
\usage{
probs_function(theta, observed_muts, t, s, states, states_matrix)
}
\arguments{
\item{theta}{(s(s+1)/2)x1 vector, where first s entries are single cut rates
(the diagonal of theta matrix), the next s-1 entries are the
2 site mutation rates (off diagonal), and so on.}

\item{observed_muts}{Observed mutations matrix
(output of observed_mutations2$obsM))}

\item{t}{Time duration}

\item{s}{Number of sites}

\item{states}{List of states (output of state_space_list)}

\item{states_matrix}{Matrix of states (output of state_space_matrix)}
}
\value{
An sxs matrix with (i, i) the probability of a single cut at site i,
         and (i, j), i < j, the probability of an overlapping cut from
         sites i to j.
}
\description{
This function computes the probability of observing a mutation of a given
type given the rate vector theta and time t.
}
