% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal_times.R
\name{proj_l1}
\alias{proj_l1}
\title{Project a Vector onto the L1 Ball}
\usage{
proj_l1(v, a)
}
\arguments{
\item{v}{A numeric vector.}

\item{a}{The radius (> 0) of the L1 ball.}
}
\value{
The projection of \code{v} onto the L1 ball of radius \code{a}.
}
\description{
Computes the Euclidean projection of a vector onto the L1 ball of a given
radius. This is used to keep the vector of intercoalescent times inside the
constraint set \eqn{\sum_i u_i \le T}, where \eqn{T} is the height of the
tree.
}
\details{
Adapted from \url{https://rdrr.io/github/vguillemot/sparseMCA/src/R/projl1.R}
}
\examples{
proj_l1(1:10, 21)

}
