% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blt.R
\name{read_blt}
\alias{read_blt}
\title{read_blt Function}
\usage{
read_blt(filepath, weight = F)
}
\arguments{
\item{filepath}{Filepath of blt file, including filename.blt}

\item{weight}{Optional parameter to return ballot weight. Default = F}
}
\description{
Function to parse and process OpenSTV / OpaVote .blt files into a format with 1 ballot per line and the candidates encoded as factors.
}
\examples{
read_blt("C:/Users/User1/Documents/leader.blt")
}
