% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{relabel}
\alias{relabel}
\title{Relabel States}
\usage{
relabel(states)
}
\arguments{
\item{(states)}{A list of states}
}
\value{
A matrix of relabeled states
}
\description{
Helper function for state_space and state_space_matrix.
Renames overlapping mutations to (start + 1) where start
is the leftmost index. For example, state_space_list(3)
will return a list containing states 122 and 022 that will be
relabeled to 133 and 033 respectively.
}
\examples{
relabel([[111],[011],[221],[101],[001],[122],[022],[222],[110],[010],[220],[100],[000]])
#       [,1] [,2] [,3]
#  [1,]    1    1    1
#  [2,]    0    1    1
#  [3,]    2    2    1
#  [4,]    1    0    1
#  [5,]    0    0    1
#  [6,]    1    3    3
#  [7,]    0    3    3
#  [8,]    2    2    2
#  [9,]    1    1    0
# [10,]    0    1    0
# [11,]    2    2    0
# [12,]    1    0    0
# [13,]    0    0    0
}
