% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_data_finite_alleles}
\alias{simulate_data_finite_alleles}
\title{Simulate Data Finite Alleles}
\usage{
simulate_data_finite_alleles(tree, Q, states_matrix, M)
}
\arguments{
\item{(tree)}{An ape tree object}

\item{(Q)}{Rate matrix for mutation states (output of Q_matrix)}

\item{(states)}{Matrix storing possible mutation states (output of state_space_matrix)
M(i, i) is the number of alleles at site i, M(i, j) is the number of alleles from a simultaneous cut
at sites i, j. In principle these numbers could be the same and we could input a single parameter,
but we will leave it as matrix for full generality.}

\item{(M)}{Symmetric matrix which stores the number of alleles per site or site group}
}
\value{
A list where the first entry is the matrix of mutation states
for the nodes and second entry is matrix of allele states.
}
\description{
This function simulates allele data for all nodes of the tree
(the leaves are the first n rows), by starting with the state (0, 0, ...0)
at the root and simulating the Markov mutation process down the tree.
}
\examples{
simul1<-coalsim(samp_times = 0, n_sampled = 5, traj = exp_traj,method="tt",val_upper=11)
mytree <- generate_newick((simul1))$newick
# Phylogenetic tree with 5 tips and 4 internal nodes.
#
# Tip labels:
#   t2_0, t3_0, t4_0, t5_0, t1_0
#
#  Rooted; includes branch lengths.

theta = matrix(c(1,1,1,0,1,1,0,0,1), nrow=3, byrow=TRUE)
Q = Q_matrix(states_matrix, theta)
#       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#  [1,]    0    0  0.0    0  0.0  0.0  0.0  0.0    0   0.0   0.0   0.0   0.0
#  [2,]    1   -1  0.0    0  0.0  0.0  0.0  0.0    0   0.0   0.0   0.0   0.0
#  [3,]    0    0  0.0    0  0.0  0.0  0.0  0.0    0   0.0   0.0   0.0   0.0
#  [4,]    1    0  0.0   -1  0.0  0.0  0.0  0.0    0   0.0   0.0   0.0   0.0
#  [5,]    0    1  0.5    1 -2.5  0.0  0.0  0.0    0   0.0   0.0   0.0   0.0
#  [6,]    0    0  0.0    0  0.0  0.0  0.0  0.0    0   0.0   0.0   0.0   0.0
#  [7,]    0    0  0.0    0  0.0  1.0 -1.0  0.0    0   0.0   0.0   0.0   0.0
#  [8,]    0    0  0.0    0  0.0  0.0  0.0  0.0    0   0.0   0.0   0.0   0.0
#  [9,]    1    0  0.0    0  0.0  0.0  0.0  0.0   -1   0.0   0.0   0.0   0.0
# [10,]    0    1  0.0    0  0.0  0.0  0.0  0.2    1  -2.2   0.0   0.0   0.0
# [11,]    0    0  1.0    0  0.0  0.0  0.0  0.0    0   0.0  -1.0   0.0   0.0
# [12,]    0    0  0.0    1  0.0  0.5  0.0  0.0    1   0.0   0.0  -2.5   0.0
# [13,]    0    0  0.0    0  1.0  0.0  0.5  0.2    0   1.0   0.5   1.0  -4.2

states_matrix = state_space_matrix(S)
#       [,1] [,2] [,3]
#  [1,]    1    1    1
#  [2,]    0    1    1
#  [3,]    2    2    1
#  [4,]    1    0    1
#  [5,]    0    0    1
#  [6,]    1    3    3
#  [7,]    0    3    3
#  [8,]    2    2    2
#  [9,]    1    1    0
# [10,]    0    1    0
# [11,]    2    2    0
# [12,]    1    0    0
# [13,]    0    0    0

M = matrix(20, nrow = S, ncol = S)
simulate_data_finite_alleles(mytree, Q, states_matrix, M)
# [[1]]
#      [,1] [,2] [,3]
# [1,]    2    2    1
# [2,]    0    0    1
# [3,]    1    1    0
# [4,]    1    1    0
# [5,]    0    0    1
# [6,]    0    0    0
# [7,]    0    0    0
# [8,]    0    0    0
# [9,]    1    1    0
#
# [[2]]
#      [,1]    [,2]    [,3]
# [1,] "12:17" "12:17" "3:14"
# [2,] "0"     "0"     "3:16"
# [3,] "1:5"   "2:5"   "0"
# [4,] "1:5"   "2:5"   "0"
# [5,] "0"     "0"     "3:13"
# [6,] "0"     "0"     "0"
# [7,] "0"     "0"     "0"
# [8,] "0"     "0"     "0"
# [9,] "1:5"   "2:5"   "0"

}
