% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{site_alleles}
\alias{site_alleles}
\title{Generate Site Allele Labels}
\usage{
site_alleles(s, num)
}
\arguments{
\item{s}{A character or numeric identifier for the site.}

\item{num}{Integer giving the number of alleles to generate.}
}
\value{
A character vector of length `num` with elements
`"s:1"`, `"s:2"`, ..., `"s:num"`.
}
\description{
Creates a vector of allele labels of the form `"s:j"` for
indices from 1 through `num`.
}
\examples{
site_alleles(3, 3)
# c("3:1", "3:2", "3:3")

}
