% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{state_space}
\alias{state_space}
\title{State Space Hash Table}
\usage{
state_space(S)
}
\arguments{
\item{(s)}{Number of sites}
}
\value{
A hash table mapping state strings to their indices
}
\description{
This generates the state space hash table for a given number of sites
}
\examples{
state_space(3)
# <hash> containing 13 key-value pair(s).
# 000 : 13
# 001 : 5
# 010 : 10
# 011 : 2
# 033 : 7
# 100 : 12
# 101 : 4
# 110 : 9
# 111 : 1
# 133 : 6
# 220 : 11
# 221 : 3
# 222 : 8

}
