% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{state_space_list}
\alias{state_space_list}
\title{State Space List}
\usage{
state_space_list(S)
}
\arguments{
\item{(S)}{}
}
\value{
A list of all possible states
}
\description{
The state space contains all possible vectors of length S,
where 0 indicates no mutation, 1 indicates mutation just
at that site, and larger numbers i are mutations that are
shared between multiple adjacent sites
}
\examples{
state_space_list(3)
# [[111],[011],[221],[101],[001],[122],[022],[222],[110],[010],[220],[100],[000]]
}
