% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{state_space_matrix}
\alias{state_space_matrix}
\title{State Space Matrix}
\usage{
state_space_matrix(S)
}
\arguments{
\item{(S)}{Number of sites}
}
\value{
An nxS matrix where each row represents a state
with one column for each site.
}
\description{
Turns the state space list into a matrix and relabels
}
\examples{

state_space_matrix(3)
#       [,1] [,2] [,3]
#  [1,]    1    1    1
#  [2,]    0    1    1
#  [3,]    2    2    1
#  [4,]    1    0    1
#  [5,]    0    0    1
#  [6,]    1    3    3
#  [7,]    0    3    3
#  [8,]    2    2    2
#  [9,]    1    1    0
# [10,]    0    1    0
# [11,]    2    2    0
# [12,]    1    0    0
# [13,]    0    0    0

}
