% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tomatoes}
\alias{tomatoes}
\title{FRED Tomato data}
\format{
\subsection{\code{lettuce}}{

A data frame with 119 rows and 5 columns:
\describe{
\item{date}{a Date object with the data of the data}
\item{series_id}{APU0000712311, tomato series}
\item{value}{the average price of tomatoes}
\item{realtime_start, realtime_end}{The date the data was downloaded}
}
}
}
\source{
\url{https://fred.stlouisfed.org/series/APU0000712311}
}
\usage{
tomatoes
}
\description{
A subset of the time series APU0000712311 from FRED, data on tomato
prices from 2016 to 2026
}
\keyword{datasets}
