% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{transition2}
\alias{transition2}
\title{Transition Function}
\usage{
transition2(parent, t, q, states)
}
\arguments{
\item{parent}{Parent state vector}

\item{t}{Time duration}

\item{q}{Rate matrix (output from q_matrix)}

\item{states}{Matrix of possible states (output from state_space_matrix)}
}
\value{
A transition in time t from a given parent and rate matrix q
}
\description{
Computes the transition probabilities given a parent state, time t,
and rate matrix q, and samples a new state.
}
