% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_theta.R
\name{unflatten_upper}
\alias{unflatten_upper}
\title{Reconstruct Upper Triangular Matrix from Flattened Vector}
\usage{
unflatten_upper(v)
}
\arguments{
\item{v}{A numeric vector containing the upper triangular elements, ordered
by diagonal (as returned by \code{flatten_upper()}).}
}
\value{
A square matrix \code{M} such that \code{flatten_upper(M) == v}.
}
\description{
Converts a vector created by \code{flatten_upper()} back into a square
upper triangular matrix (including main diagonal). Zeros are filled in
the lower triangular part.
}
\examples{
theta <- c(1,2,3,4,5,6)
unflatten_upper(theta)
#      [,1] [,2] [,3]
# [1,]    1    4    6
# [2,]    0    2    5
# [3,]    0    0    3

}
