% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR1.R
\name{AR1}
\alias{AR1}
\title{First-order autoregressive model}
\usage{
AR1(x)
}
\arguments{
\item{x}{a vector, ideally a timeseries vector}
}
\value{
an Arima model
}
\description{
This function is a wrapper around the forecast::Arima() function.
It creates a first-order autoregressive model with a constant term.
}
