% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLT-package.R
\docType{package}
\name{BLT-package}
\alias{BLT}
\alias{BLT-package}
\title{BLT: Time Series Helpers for Working with BLT Data}
\description{
This is a demo package for the "Code you can bank on" workshop at the Federal Reserve. Helper functions for working with time series data, especially data about bacon, lettuce, and tomatoes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AmeliaMN/BLT}
  \item \url{https://ameliamn.github.io/BLT/}
  \item Report bugs at \url{https://github.com/AmeliaMN/BLT/issues}
}

}
\author{
\strong{Maintainer}: Amelia McNamara \email{amelia.mcnamara@stthomas.edu}

}
\keyword{internal}
