% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLT_ggplot_na_distribution.R
\name{BLT_ggplot_na_distribution}
\alias{BLT_ggplot_na_distribution}
\title{Wrapper around ggplot_na_distribution with aesthetics}
\usage{
BLT_ggplot_na_distribution(data, mapping = aes(), ...)
}
\arguments{
\item{data}{Default dataset to use for plot}

\item{mapping}{set of aesthetic mappings created by aes()

ggplot_na_distribution() understands the x aesthetic.}
}
\value{
a ggplot object
}
\description{
Wrapper around imputeTS::ggplot_na_distribution that takes
data and mapping arguments like the rest of ggplot2
}
\examples{
bacon |>
BLT_ggplot_na_distribution(mapping = ggplot2::aes(x = value))
BLT_ggplot_na_distribution(bacon, mapping = ggplot2::aes(x = value))
}
