% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_missing_tidy.R
\name{perc_missing_tidy}
\alias{perc_missing_tidy}
\title{A tidy version of perc_missing}
\usage{
perc_missing_tidy(dataset, variable)
}
\arguments{
\item{dataset}{a dataset}

\item{variable}{a variable within that dataset, ideally a time series}
}
\value{
a tibble containing the percent of the variable that was missing
}
\description{
A wrapper around \link[imputeTS:statsNA]{imputeTS::statsNA} that accepts tidy arguments
}
\examples{
bacon |> perc_missing_tidy(value)

lettuce |> perc_missing_tidy(value)
}
