% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{probs_function}
\alias{probs_function}
\title{Probability of Mutations}
\usage{
probs_function(theta, observedMuts, t, S, states, states_matrix)
}
\arguments{
\item{(theta)}{(S(S+1)/2)x1 vector, where first S entries are single cut rates
(the diagonal of theta matrix), the next S-1 entries are the
2 site mutation rates (off diagonal), and so on.}

\item{(observedMuts)}{Observed mutations matrix (output of observed_mutations2$obsM))}

\item{(t)}{Time duration}

\item{(S)}{Number of sites}

\item{(states)}{List of states (output of state_space_list)}

\item{(states_matrix)}{Matrix of states (output of state_space_matrix)}
}
\value{
An SxS matrix with (i, i) the probability of a single cut at site i,
         and (i, j), i < j, the probability of an overlapping cut from sites i to j.
}
\description{
This function computes the probability of observing a mutation of a given
type given the rate vector theta and time t.
}
