% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Q_matrix}
\alias{Q_matrix}
\title{Q Matrix}
\usage{
Q_matrix(states, theta)
}
\arguments{
\item{(states)}{A matrix of states (output of state_space_matrix) with length n}

\item{(theta)}{An SxS matrix of cut rates}
}
\value{
A nxn generator matarix
}
\description{
Q is the generator matrix for a continuous time Markov chain over the list
of states. Theta is an SxS matrix with theta[i, j] is the rate of a cut from
site i to site j. Theta[i, i] is the rate of a single cut at site i.
}
